fixup! Added docker setup for running xdebug

+++ xdebug/internal_install_xdebug.sh
@@ -2,25 +2,36 @@
 
 XDEBUG_VERSION=3.2.2
 
+option=$1
+forceReinstall=""
+
+if [[ "$option" = "--force-reinstall" ]]; then
+    forceReinstall="true"
+fi
 
 if [ ! -f /usr/local/etc/php/conf.d/99-xdebug.ini ]; then
-    echo "Installing XDEBUG"
-    rm -f /tmp/xdebug.tar.gz xdebug-${XDEBUG_VERSION}
-    curl https://xdebug.org/files/xdebug-${XDEBUG_VERSION}.tgz --output /tmp/xdebug.tar.gz
-    # apt-get install php-dev autoconf automake
-    apt-get update
-    apt-get install -y autoconf automake libtool iputils-ping telnet procps
+    if [ ! -f /usr/local/lib/php/extensions/no-debug-non-zts-20210902/xdebug.so ] || [ "$forceReinstall" = "true" ]; then
+        echo "(Re)Installing XDEBUG"
+        rm -f /tmp/xdebug.tar.gz xdebug-${XDEBUG_VERSION}
+        curl https://xdebug.org/files/xdebug-${XDEBUG_VERSION}.tgz --output /tmp/xdebug.tar.gz
+        # apt-get install php-dev autoconf automake
+        apt-get update
+        apt-get install -y autoconf automake libtool iputils-ping telnet procps
 
 
-    cd /tmp
-    tar -xzf xdebug.tar.gz
-    cd xdebug-${XDEBUG_VERSION}
-    phpize
-    # autoreconf -i
-    ./configure
-    make
+        cd /tmp
+        tar -xzf xdebug.tar.gz
+        cd xdebug-${XDEBUG_VERSION}
+        phpize
+        # autoreconf -i
+        ./configure
+        make
 
-    cp modules/xdebug.so /usr/local/lib/php/extensions/no-debug-non-zts-20210902
+        cp modules/xdebug.so /usr/local/lib/php/extensions/no-debug-non-zts-20210902
+
+        # Install validator scripts ( see Settings -> PHP -> Debug -> Validate debugger configuration on the Web Server -> Debug validaton script
+        sudo -u www-data bash -c "cd /var/www/public && curl -f -L  -o ./phpstorm_xdebug.zip "https://packages.jetbrains.team/files/p/ij/xdebug-validation-script/script/phpstorm_xdebug_validator.zip" && unzip -n ./phpstorm_xdebug.zip -d . && rm -f ./phpstorm_xdebug.zip"
+    fi
 
     echo "zend_extension = xdebug.so" > /usr/local/etc/php/conf.d/99-xdebug.ini
     #  echo "xdebug.mode=develop,coverage,debug,profile" >> /usr/local/etc/php/conf.d/99-xdebug.ini
@@ -32,9 +43,6 @@ if [ ! -f /usr/local/etc/php/conf.d/99-xdebug.ini ]; then
     echo "xdebug.client_port=$XDEBUG_PORT" >> /usr/local/etc/php/conf.d/99-xdebug.ini
     echo "xdebug.client_host=host.docker.internal" >> /usr/local/etc/php/conf.d/99-xdebug.ini
     echo "" >> /usr/local/etc/php/conf.d/99-xdebug.ini
-
-    # Install validator scripts ( see Settings -> PHP -> Debug -> Validate debugger configuration on the Web Server -> Debug validaton script
-    sudo -u www-data bash -c "cd /var/www/public && curl -f -L  -o ./phpstorm_xdebug.zip "https://packages.jetbrains.team/files/p/ij/xdebug-validation-script/script/phpstorm_xdebug_validator.zip" && unzip -n ./phpstorm_xdebug.zip -d . && rm -f ./phpstorm_xdebug.zip"
 
 else
     echo "XDEBUG already installed !"
